Framework updated V1.0
