+++ solutions/domain-specific-language/src/main/java/com/hansencx/solutions/portal/PortalPageGenerator.java
@@ -1,0 +1,12 @@
+package com.hansencx.solutions.portal;
+
+/**
+ * @param
+ * @author Nhi Dinh
+ * @return
+ * @since 1/16/2019
+ */
+
+
+public class PortalPageGenerator {
+}
