--- solutions/framework/src/main/java/com/hansencx/solutions/config/extent-config.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<extentreports>
-    <configuration>
-        <!-- report theme -->
-        <!-- standard, dark -->
-        <theme>standard</theme>
-    
-        <!-- document encoding -->
-        <!-- defaults to UTF-8 -->
-        <encoding>UTF-8</encoding>
-        
-        <!-- protocol for script and stylesheets -->
-        <!-- defaults to https -->
-        <protocol>https</protocol>
-        
-        <!-- title of the document -->
-        <documentTitle>Extent</documentTitle>
-        
-        <!-- report name - displayed at top-nav -->
-        <reportName>Automation Report</reportName>
-        
-        <!-- location of charts in the test view -->
-        <!-- top, bottom -->
-        <testViewChartLocation>top</testViewChartLocation>
-		
-        <!-- custom javascript -->
-        <scripts>
-            <![CDATA[
-                $(document).ready(function() {
-                    
-                });
-            ]]>
-        </scripts>
-        
-        <!-- custom styles -->
-        <styles>
-            <![CDATA[
-                
-            ]]>
-        </styles>
-    </configuration>
-</extentreports>
