--- solutions/framework/src/main/java/utilities/logger/log4j.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/" debug="true">
-
-    <appender name="fileAppender" class="org.apache.log4j.FileAppender">
-        <param name="Threshold" value="INFO" />
-        <param name="File" value="${logLocation}"/>
-        <param name="immediateFlush" value="true"/>
-        <param name="threshold" value="debug"/>
-        <param name="append" value="true"/>
-
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d %-5p [%c{1}] %m %n" />
-        </layout>
-    </appender>
-       <root>
-           <level value="INFO"/>
-        <appender-ref ref="fileAppender"/>
-    </root>
-
-</log4j:configuration>
