--- solutions/page-objects/src/main/java/com/hansencx/solutions/pages/PostCategoriesPage.java
@@ -1,9 +1,0 @@
-package com.hansencx.solutions.pages;
-
-import com.hansencx.solutions.core.BasePage;
-import org.openqa.selenium.WebDriver;
-
-public class PostCategoriesPage extends BasePage {
-    public PostCategoriesPage(WebDriver driver){super(driver);}
-
-}
