--- solutions/page-objects/src/main/java/com/hansencx/solutions/pages/portal/HomePage.java
@@ -1,11 +1,0 @@
-package com.hansencx.solutions.pages.portal;
-
-import com.hansencx.solutions.core.BasePage;
-import org.openqa.selenium.WebDriver;
-
-public class HomePage extends BasePage {
-    public HomePage(WebDriver driver){
-        super(driver);
-    }
-
-}
