--- solutions/page-objects/src/main/java/com/hansencx/solutions/pages/portal/utilities/Links.java
@@ -1,14 +1,0 @@
-package com.hansencx.solutions.pages.portal.utilities;
-
-/**
- * @param
- * @author Nhi Dinh
- * @return
- * @since 1/15/2019
- */
-
-
-public class Links {
-    public static final String DOMAIN= "https://solutionsportalqa.hansencx.com";
-    public static final String URL_LOGIN = DOMAIN+"/Portal/Login.aspx ";
-}
