--- solutions/page-objects/src/main/java/com/hansencx/solutions/pages/portal/utilities/Portal_PageGenerator.java
@@ -1,51 +1,0 @@
-package com.hansencx.solutions.pages.portal.utilities;
-
-import com.hansencx.solutions.core.BasePage;
-import com.hansencx.solutions.pages.portal.HomePage;
-import com.hansencx.solutions.pages.portal.LoginPage;
-import com.hansencx.solutions.pages.portal.SearchPage;
-import com.hansencx.solutions.pages.portal.SearchResultPage;
-import com.hansencx.solutions.pages.portal.navigation.LeftNavigation;
-import com.hansencx.solutions.pages.portal.navigation.TopNavigation;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.PageFactory;
-
-/**
- * @param
- * @author Nhi Dinh
- * @return
- * @since 1/15/2019
- */
-
-
-public class Portal_PageGenerator {
-    public static WebDriver driver;
-
-    public Portal_PageGenerator(WebDriver driver){
-        this.driver = driver;
-    }
-
-    private <TPage extends BasePage> TPage GetPage(Class<TPage> pageClass){
-        return PageFactory.initElements(driver, pageClass);
-    }
-
-    public LoginPage Login() {
-        return GetPage(LoginPage.class);
-    }
-
-    public HomePage Home(){
-        return GetPage(HomePage.class);
-    }
-    public SearchPage Search(){
-        return GetPage(SearchPage.class);
-    }
-    public SearchResultPage SearchResult(){
-        return GetPage(SearchResultPage.class);
-    }
-    public TopNavigation TopNavigation(){
-        return GetPage(TopNavigation.class);
-    }
-    public LeftNavigation LeftNavigation(){
-        return GetPage(LeftNavigation.class);
-    }
-}
