+++ solutions/page-objects/src/main/java/com/hansencx/solutions/portal/pages/HomePage.java
@@ -1,0 +1,11 @@
+package com.hansencx.solutions.portal.pages;
+
+import com.hansencx.solutions.core.BasePage;
+import org.openqa.selenium.WebDriver;
+
+public class HomePage extends BasePage {
+    public HomePage(WebDriver driver){
+        super(driver);
+    }
+
+}
