+++ solutions/page-objects/src/main/java/com/hansencx/solutions/portal/pages/SearchResultPage.java
@@ -1,0 +1,21 @@
+package com.hansencx.solutions.portal.pages;
+
+import com.hansencx.solutions.core.BasePage;
+import org.openqa.selenium.WebDriver;
+
+/**
+ * @param
+ * @author Nhi Dinh
+ * @return
+ * @since 1/15/2019
+ */
+
+
+public class SearchResultPage extends BasePage {
+    public SearchResultPage(WebDriver driver){
+        super(driver);
+    }
+
+
+
+}
